==> code/3_analysis/scratch_from_all_steps/4_study_ag&counterfactual.Rmd <==
---
title: "4_study_ag&counterfactual"
author: "Anna Boser"
date: "10/4/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(sf)
library(raster)
library(here)
library(tmap)
library(stringr)
library(dplyr)
library(stars)
```

This is supposed to generate the ag and counterfactual shapefiles (2019 and 2020) to send to apppeears. 
To do so, I:
1) Compare the 2018 CDL and DWR datasets to assure myself that the CDL will give me a good ag boundary (I can also add USGS comparison in here for fun)
2) Create shapefiles of all ag areas in 2019 and 2020
3) Make shapefiles of all counterfactual land cover types in 2019 and 2020 for a certain buffer around all ag regions

## 1: Compare the 2018 CDL and DWR datasets to assure myself that the CDL will give me a good ag boundary (I can also add USGS comparison in here for fun)

### read in data
```{r}
# CDL (2018)
CDL <- raster(here("data", "raw", "CDL", "CDL2018", "2018_30m_cdls.img"))

# DWR
DWR <- read_sf(here("data", "raw", "DWR_crop", "i15_Crop_Mapping_2018.shp")) %>%
  st_transform(crs(CDL))
DWR <- st_zm(DWR) # DWR in 3 dims with 0 for z value

# CA counties for cropping
counties <- st_read(here("data", "raw", "shapefiles", "study_area_counties", "study_area_counties.shp")) %>%
  st_transform(crs(CDL))

CA <- st_as_sf(st_union(counties))

# crop the CDL to California
CDL <- crop(CDL, counties)
```


```{r}
#small example for debugging
# CDL <- crop(CDL, DWR[1:10,])
# DWR <- DWR[1:10,]
```

What I want is a confusion matrix of ag land and non ag land in California. 
So I need to create two rasters with only 0 and 1 values for ag and non-ag, and compare them. 

### create 0 and 1 raster for CDL
```{r}
code_dictionary <- read.csv(file = here("data", 
                      "intermediate", 
                      "CDL_code_dictionary.csv"))

# make a dictionary for converting CDL numbers to cultivated and uncultivated. 
code_dictionary$cultivated <- ifelse(code_dictionary$cultivated == "Cultivated", 1, 0)
cultivate_dic <- code_dictionary$cultivated
names(cultivate_dic) <- code_dictionary$code

CDL_cultivated <- CDL
values(CDL_cultivated) <- cultivate_dic[as.character(values(CDL_cultivated))]
```

### create a 0 and 1 raster for DWR
```{r}
DWR_cultivated <- rasterize(DWR, CDL, field = 1, fun = "max")
values(DWR_cultivated) <- ifelse(is.na(values(DWR_cultivated)), 0, 1)
```

## create the confusion matrix
```{r}
DWR_CDL <- sum(values(DWR_cultivated)*values(CDL_cultivated))
notDWR_notCDL <- sum(ifelse(values(DWR_cultivated) == 0, 1, 0)*ifelse(values(CDL_cultivated) == 0, 1, 0))
DWR_notCDL <- sum(values(DWR_cultivated)*ifelse(values(CDL_cultivated) == 0, 1, 0))
notDWR_CDL <- sum(ifelse(values(DWR_cultivated) == 0, 1, 0)*values(CDL_cultivated))

confusion <- data.frame("DWR" = c(DWR_CDL, DWR_notCDL), "not_DWR" = c(notDWR_CDL, notDWR_notCDL), "rownames" = c("CDL", "not_CDL"), row.names = "rownames")

confusion$Sum <- apply(confusion, 1, sum)
confusion["Sum",] <- apply(confusion, 2, sum)
```


## 2) Create shapefiles of all ag areas in 2019 and 2020
```{r}
DWR_flat <- st_as_sf(st_union(DWR))
```


## 3) Make shapefiles of all counterfactual land cover types in 2019 and 2020 for a certain buffer around all ag regions

```{r}
DWR_buffer <- st_buffer(DWR_flat, 20000) # 20 km buffer. Use st_crs(DWR_flat) to see that the unit for this projection is meters (also checked it makes sense in google maps)

# Only keep what's in the buffer but not Ag according to DWR and only 100% the counterfactual in CDL

# placeholder for removing ag

# make a raster with 0 and 1 where there is the counterfactual present
# create dictionary
counter_dic <- code_dictionary$counterfactual
names(counter_dic) <- code_dictionary$code

CDL_counter <- crop(CDL, DWR_buffer)
CDL_counter <- mask(CDL_counter, DWR_buffer)
values(CDL_counter) <- counter_dic[as.character(values(CDL_counter))]
values(CDL_counter) <- ifelse(values(CDL_counter) == 1, 1, NA)

counterfactual_sf <- st_as_sf(st_as_stars(CDL_counter), as_points = FALSE, merge = TRUE) %>% st_union() %>% st_as_sf()
```
