==> code/3_analysis/scratch_from_all_steps/6.75_PET_resample.ipynb <==
from osgeo import gdal, gdalconst
from pyprojroot import here

PET_filename = str(here("./data/intermediate/PET/PET_rolling_avg_OGres.tif"))
PET = gdal.Open(src_filename, gdalconst.GA_ReadOnly)
PET_proj = PET.GetProjection()
PET_geotrans = PET.GetGeoTransform()

band = PET.GetRasterBand(1)
array = band.ReadAsArray()


here()
